Fix level and score functions

+++ core/src/io/intino/gamification/core/box/events/mission/MissionDifficulty.java
@@ -3,13 +3,13 @@ package io.intino.gamification.core.box.events.mission;
 public enum MissionDifficulty {
     Easy(1), Medium(2), Hard(3);
 
-    private final int multiplier;
+    private final float multiplier;
 
-    MissionDifficulty(int multiplier) {
+    MissionDifficulty(float multiplier) {
         this.multiplier = multiplier;
     }
 
-    public int multiplier() {
+    public float multiplier() {
         return multiplier;
     }
 }
